--- src/server.js
@@ -1,3 +1,0 @@
-/**
- * This is the non gulp / stream version of the server 
- */
